style : 공백 등 정리
