+++ src/main/java/racingcar/CarList.java
@@ -3,8 +3,6 @@ package racingcar;
 import java.util.ArrayList;
 import java.util.stream.Stream;
 
-import camp.nextstep.edu.missionutils.Console;
-
 public class CarList {
 	private ArrayList<Car> carList;
 	private ArrayList<String> carNameList;
@@ -92,28 +90,6 @@ public class CarList {
 			String winners = String.join(", ", winnerList);
 			System.out.printf("최종 우승자 : %s", winners);
 		}
-
-	}
-
-}
-
-// 문자열을 입력받아서, 스트림으로 구분해서 반환한다.
-class StreamMaker {
-	private String inputStr;
-
-	Stream<String> getCarNames() {
-		return Stream.of(inputStr.split(","));
-	}
-
-	StreamMaker() {
-		inputStrOfCarNames();
-	}
-
-	private void inputStrOfCarNames() {
-		inputStr = Console.readLine();
-
-		if (!inputStr.matches("[a-zA-Z0-9,]+")) {
-			throw new IllegalArgumentException("[ERROR] 자동차이름(영문/숫자), 쉼표만 입력 가능합니다.");
-		}
 	}
 }
+
